Remove checks for (mostly) standard C functions
This commit removes checks for functions and headers mostly required to
exist in a standard C (or C++) environment such as the `stdlib.h` header
or various string-related functions.

Some of the checks were looking for functions that are non-standard, but
the function is not (or no longer) used in UG, for example the test for
the (deprecated) `bzero()` function. These have also been removed.