Switch versioning scheme from UG versioning to Dune core versioning
Up to now, the version of the dune-uggrid module was 3.13.0, because that
was the version of UG when we forked it.  However, since dune-uggrid and
dune-grid are not separated by official interfaces, the two modules
practically have to be developed in lock-step.  Therefore starting with
this patch the dune-uggrid module will follow the release cycle and
the version numbering of the Dune core modules.