Read symbol table index using the correct integer type
The function `XferPackSingleMsg` writes the index using the expression

```c++
  *(theSymTab[mi].adr.ref) = (DDD_OBJ)(mi+1);
```

that is as an integer cast to a pointer.

If this value is read on an architecture where pointers are 64bit and
INT is 32bit using the expression

```c++
  stIdx = (*(INT *)(msgrefarray+l)) - 1;
```

only 32bits of the actual value are read.  On little endian
architectures we still read the relevant bits, however on big endian
architectures this reads the upper 32bits which are 0 (the values are
small) and the relevant 32bits are discarded.

This should fix the test failures in dune-grid on [s390x] and some of
the errors on [sparc64].

  [s390x]: https://buildd.debian.org/status/fetch.php?pkg=dune-grid&arch=s390x&ver=2.5.0%7Erc1-1&stamp=1480658113
  [sparc64]: https://buildd.debian.org/status/fetch.php?pkg=dune-grid&arch=sparc64&ver=2.5.0%7Erc1-1&stamp=1480644763