Remove domain midpoint, radius, and convexity
UG used to store the midpoint and radius of a circle that contained
the domain.  This was mainly used by the graphics subsystem, which
is gone now.  Therefore we can remove the midPoint, radius, and
convexity data members, and make the code a tiny bit simpler.

The patch changes the interface of the CreateDomain method, which
is used in dune-grid.  Therefore, dune-grid needs to be updated
in lock-step with this change.  I will post the corresponding
patch shortly.