[bugfix] fix build without MPI
When MPI is not found, UG_ENABLE_PARALLEL is not defined in
CMakeLists.txt and thus, the preprocessor variable ModelP is
not defined. This in turn means that the message_buffer and
message_buffer_size member variables will not be defined which leads
to compilation errors in the accessor methods that use these member
variables.

To prevent these compilation errors when not using MPI, I moved the
accessor methods into #ifdef ModelP blocks.

(cherry picked from commit 3bced266592f6234f1de684a1233b36e55fb4fbc)