DDD: initialize counters before using them
The counters might be used by the various initialization functions and
therefore have to be set to zero earlier.

This worked previously as variables with static storage duration are
zero-initialized and DDD left them at zero when exiting as well.
However the context object's member variable is not zero-initialized
so this broke with the change as at least `nCplItems` was used by
`ddd_IFInit`.