do not sort messages by size before sending
Sorting the forward list of messages means a new first item is chosen,
however the loop to free all XFERMSG objects in DDD_XferEnd will start
from the old list head and no longer frees all entries.  This was not
noticed before the previous commit as the free in DDD_XferEnd was a
no-op and memory would be released by the heap structure instead.

Alternatively one could return the new list head (or pass in a
modifiable reference), but this seems more complex then just skipping
the sorting step.  I doubt handling low resource environments in this
way is still necessary.