DDD: xfer: add parts of state to context object
This commit moves the state from `XFER_GLOBALS` to the context
object.  Some global variables still remain, for example the lists
managed by sll.h and `theXIAddData` from cmds.cc.

The casts are necessary as the structs cannot be defined in the `xfer`
module as long as it is still dim-dependent (same as for the previous
change to the `join` module).