Rentroduce Vector::skip as Vector::leafIndex
The 'skip' data member of the Vector class used to be abused for
the leaf index of the vector, if that vector corresponds to an
element facet in 3d.  By accident, that 'skip' field got removed
in a recent commit.  Let's put it back in, but rename it to
'leafIndex' now.  All it is used for nowadays is that leaf index.