[bugfix] Fix bisection loadbalancing algorithm
ebf74cbf replaces
three sort functions by a template. However, the
template arguments where never used in the function,
so we only sorted in one coordinate direction.
This commit fixes this by replacing the hard-coded
indices with the template argument values.

Thanks to Leopold Stadler for finding the offending commit.