Make DDD_TypeDefine accept member offsets rather than member pointers
The DDD_TypeDefine method previously required pointers to the individual
data members of a data structure to compute its layout.  For this you
need a pointer to the beginning of the data structure, and the code
used stuff like nullptr and pointers onto the stack for this.
This worked (mostly), but working with objects starting at 0
involves undefined behavior.

This patch makes the code more robust by using the standard method
'offsetof' to compute the data structure layouts.  This is done
outside of DDD_TypeDefine, which hence changes its interface.
This is not much of a problem, because the method is only used
in initddd.cc anyway.