Merge the PreInitElementTypes method and the InitElementTypes method
Previously, PreInitElementTypes would initialize that part of the
element descriptions that did not depend on the 'format', i.e.,
the layout of simulation data.  It was called right at program
startup.  Then, the user would provide a 'format', i.e., a choice
of simulation data variables.  Only after that was known the
element descriptions could be completed.  For that, the
InitElementTypes method existed, and was called only after
a multigrid object had been constructed.

Since we don't really have 'format's anymore, there is no need
for the separation between the two init methods, and this patch
merges them into one.  The new method is now called at startup.
This simplifies the startup logic a little bit.