Remove macros ending in _REF
These access element information indexed by the number of element
vertices.  They are redundant, because for each macro there is a
corresponding macro ending in _TAG, which provides the same
information indexed by an (integer) tag.  (Coincidentally, the
tag *is* the number of vertices).

Remove the _REF macros to remove the redundancy.