Compute inner_size and bnd_size from actual class declarations
Previously, the size of the classes implementing the different
elements types (such as triangles and tetrahedra) where computed
in the ProcessElementDescription method. With this patch they
are now taken from the actual class declarations in gm.h.

As it turns out, the previously computed values don't actually
match the values one would expect from the class declarations.
Therefore, the code now subtracts a few bytes seemingly at
random.  I think this is needed to compensate for a bug
somewhere else.  This will have to be investigated later.