Rename the outer namespace from UG:: to Dune::UG::
Rationale:

a) This is more consistent with the rest of Dune.

b) I want to start using classes from dune-common in dune-uggrid.
   Without this change, I need to prefix them with Dune:: all the time.
   Having everything in the Dune:: namespace avoids this.

Curiously, no change to dune-grid is needed, even though the UG::
namespace name is used there.  The reason is that all dune-grid code
is in the Dune namespace, too.