Define DOUBLE_VECTOR to be a FieldVector<DOUBLE,DIM>
This patch is a first step towards replacing C arrays that represent
points by FieldVector.  The motivation is to have the compiler
enforce a bit more type safety.

This patch uses casts in a few places to go from DOUBLE[] to
FieldVector.  They will disappear in later commits when more
arrays are replaced by FieldVector.