Move the class linear_segment out of the environment data structure
Instead, store it directly in the 'domain' data structure.
This simplifies setting up the domain boundary a little bit.
Also, it saves a bit of memory, because environment items
store some extra information.  In particular, they store
a name in an array of fixed size char[128].  In the case
of linear_segment objects, however, these names are never
actually used.