Simplify how STD_BVP_Configure is called
Previously, it was a static method and the BVP_Descriptor object
held a pointer to that method.  In the old days, this was
used to make the Configure method, well, configurable.

Now we just expose the method in std_domain.h and
let the UGGrid implementation in dune-grid call the method
directly.