Allocate the right amount of memory for 'node' objects
These objects used to have a member of type VECTOR*, but whether
memory was actually allocated for that member was determined at run-time
(via the 'format', which doesn't exist anymore).  The VECTOR* member
was removed a while ago, but the run-time size computation was left
untouched.  This lead to valgrind complaints when load balancing.
This commit fixes the size computations.