Remove the BVP type
BVP was a pointer to a forward-declaration of the class
std_BoundaryValueProblem.  That's not good -- in particular,
the destructor of std_BoundaryValueProblem would not be called
when handling the class through a BVP pointer.  valgrind
would show this in certain situations. I am surprised
that this is possible and legal at all.

Use STD_BVP instead, everywhere.  That is a proper pointer.